teste-ok/requirements.txt.txt
Flask==2.0.2
Flask-JWT-Extended==4.3.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Python==3.9.6
virtualenv==20.10.0